src/Spg.AloMalo.Application/Services/PhotoUseCases/Query/GetPhotosQueryHandler.cs
using MediatR;
using Spg.AloMalo.DomainModel.Dtos;
using Spg.AloMalo.DomainModel.Interfaces.Repositories;
using Spg.AloMalo.DomainModel.Model;

namespace Spg.AloMalo.Application.Services.PhotoUseCases.Query
{
    public class GetPhotosQueryHandler : IRequestHandler<GetPhotosQueryModel, List<PhotoDto>>
    {
        private readonly IReadOnlyPhotoRepository _photoRepository;

        public GetPhotosQueryHandler(IReadOnlyPhotoRepository photoRepository)
        {
            _photoRepository = photoRepository;
        }

        public Task<List<PhotoDto>> Handle(GetPhotosQueryModel request, CancellationToken cancellationToken)
        {
            IPhotoFilterBuilder builder =
                _photoRepository
                .FilterBuilder;

            string[] filters = request.Query.Filter.Split(';');

            foreach (var filter in filters)
            {
                string[] parts = filter.Split(' ');
                if (parts.Length == 3)
                {
                    string property = parts[0];
                    string operation = parts[1];
                    string value = parts[2];
                    builder = new PhotoPropertyFilter(builder, property, operation, value).Apply();
                }
            }

            return Task.FromResult(
                builder
                .Build()
                .Select(p => p.ToDto())
                .ToList()
            );
        }
    }
}
